added loader for changing naguage
