+++ src/components/Loader/Loader.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import s from './Loader.module.scss';
+
+export const Loader = () => {
+  return (
+    <div className={s.wrapper}>
+      <div className={s.ldsRoller}>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+      </div>
+    </div>
+  );
+};
